requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
nltk
joblib
flasks